tools: adapt Changelog handling in Makefile [2/2]
due the fact google is resetting all env variables, we handle
skipping the Changelog by setting USER_BUILD_NO_CHANGELOG=1 in Makefile.
Also of this, we can't read env. variable $OUT in script directly
anymore, so we add this as an argument to the script.
Same goes for the command date, so we need to add the PATH variable.

Change-Id: Id70bd968e54084d642c7c60b32166d36001510db